+++ example_test.go
@@ -46,14 +46,13 @@ var goVer = goyek.Define(goyek.Task{
 })
 
 func Example() {
-	// Use the same output for flow and flag.
-	flag.CommandLine.SetOutput(goyek.Output())
-
 	// Set the help message.
 	usage := func() {
-		fmt.Println("Usage of build: [tasks] [flags] [--] [args]")
+		out := goyek.Output()
+		flag.CommandLine.SetOutput(out)
+		fmt.Fprintln(out, "Usage of build: [tasks] [flags] [--] [args]")
 		goyek.Print()
-		fmt.Println("Flags:")
+		fmt.Fprintln(out, "Flags:")
 		flag.PrintDefaults()
 	}